修正: 浮動小数点演算を排除し SSE レジスタエラーを解決

+++ kernel/include/vbe.h
@@ -19,6 +19,6 @@ void vbe_set_pixel(int x, int y, uint32_t color);
 
 /* Responsive Helpers */
 bool vbe_is_portrait(void);
-float vbe_get_aspect_ratio(void);
+uint32_t vbe_get_aspect_ratio(void);
 
 #endif
